backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.7.0
langchain>=0.3.25
langchain-openai>=0.1.25
langgraph>=0.0.15
typing-extensions>=4.0.0
langchain-community>=0.3.24